OPDIS : Added more doxygen comments

+++ opdis/opdis.h
@@ -31,7 +31,7 @@
  * \param i The most recently disassembled instruction.
  * \param arg Argument provided when the callback is set
  * \return 0 if disassembly should halt, nonzero (1) otherwise.
- * This function is invoked after the display callback. The default 
+ * \details This function is invoked after the display callback. The default 
  * behavior is to halt disassembly only if an invalid instruction is
  * encountered. The caller can override this function in order to 
  * specify more detailed halting conditions, e.g. to halt once a
@@ -40,17 +40,21 @@
 typedef int (*OPDIS_HANDLER) ( const opdis_insn_t * i, void * arg );
 
 /*!
- * \typedef void (*OPDIS_DISPLAY) ( const opdis_insn_t * i, void * arg );
+ * \typedef void (*OPDIS_DISPLAY) ( const opdis_insn_t * i, void * arg )
  * \ingroup configuration
  * \brief Callback used to display or store a disassembled instruction
  * \param i The most recently disassembled instruction.
  * \param arg Argument provided when the callback is set
- * This function is invoked every time an instruction is disassembled; it
- * emits the instruction to the caller of the disassembler. The default
+ * \details This function is invoked every time an instruction is disassembled;
+ * it emits the instruction to the caller of the disassembler. The default
  * display function writes to STDOUT.
  * \note The caller must copy the instruction if they are going to store it;
  *       the contents of i will be overwritten when the next instruction
  *       is disassembled.
+ * \note The default display routine writes instructions to STDOUT without
+ *       storing or sorting them. This will lead to unexpected behavior
+ *       (i.e. out-of-order instruction listings) when using 
+ *       opdis_disasm_cflow(). 
  */
 typedef void (*OPDIS_DISPLAY) ( const opdis_insn_t * i, void * arg );
 
@@ -118,12 +122,12 @@ typedef struct {
 	void * resolver_arg;
 
 	/*! \var decoder
-	 *  \brief callback which builds an opdis_insn_t from libopcodes strings
+	 *  \brief callback to build an opdis_insn_t from libopcodes strings.
 	 */
 	OPDIS_DECODER decoder;
 
 	/*! \var buf
-	 *  \brief buffer for storing libopcodes strings as they are emitted 
+	 *  \brief buffer for storing libopcodes strings as they are emitted.
 	 */
 	opdis_insn_buf_t * buf;
 
@@ -179,11 +183,12 @@ void LIBCALL opdis_term( opdis_t );
  * \fn opdis_set_defaults( opdis_t )
  * \ingroup configuration
  * \brief Initializes an opdis object to default, sane values.
- * \param o opdis object to re-initialize
+ * \param o opdis disassembler to configure.
  * \sa opdis_init opdis_set_arch opdis_set_x86_syntax opdis_set_error
  * \sa opdis_set_display opdis_set_handler opdis_set_resolver opdis_set_decoder
  * \sa opdis_set_disassembler_options
  * \note The default architecture is i386, and the default syntax is Intel.
+ *       All callbacks will be set to the default internal routines.
  */
 void LIBCALL opdis_set_defaults( opdis_t o );
 
@@ -208,9 +213,9 @@ enum opdis_x86_syntax_t {
  * \details Invokes opdist_set_arch() to force the architecture to i386,
  *          and sets the libopcodes print instruction to either
  *          print_insn_i386_intel or print_insn_i386_att.
- * \param o opdis object to configure.
+ * \param o opdis disassembler to configure.
  * \param syntax The syntax option to use.
- * \note This only applied to x86 disassemblers.
+ * \note This only applies to x86 disassemblers.
  */
 void LIBCALL opdis_set_x86_syntax( opdis_t o, opdis_x86_syntax_t syntax );
 
@@ -221,6 +226,13 @@ void LIBCALL opdis_set_x86_syntax( opdis_t o, opdis_x86_syntax_t syntax );
 /*!
  * \fn opdis_set_arch( opdis_t, enum bfd_architecture, disassembler_ftype )
  * \ingroup configuration
+ * \brief Set the architecture and disassembler routine for libopcodes
+ * \details This will set architecture of the libopcodes target to a BFD 
+ *          architecture, and specify the libopcodes print_insn routine
+ *          that will be used for disassembly.
+ * \param o opdis disassembler to configure.
+ * \param arch A valid BFD architecture from /usr/include/bfd.h .
+ * \param fn A valid libopcodes print_insn routine from /usr/include/dis-asm.h .
  */
 void LIBCALL opdis_set_arch( opdis_t o, enum bfd_architecture arch, 
 			     disassembler_ftype fn );
@@ -228,64 +240,126 @@ void LIBCALL opdis_set_arch( opdis_t o, enum bfd_architecture arch,
 /*!
  * \fn opdis_set_disassembler_options( opdis_t, const char * )
  * \ingroup configuration
+ * \brief Set libopcodes disassembler options
+ * \details Sets the libopcodes disassembler_options config field. Opdis passes
+ *          this string blindly to libopcodes. To view the valid options for
+ *          a disassembler, invoke the appropriate print_*_disassembler_options
+ *          routine listed in /usr/include/dis-asm.h .
+ * \param o opdis disassembler to configure.
+ * \param options The options string for the disassembler.
  */
 void LIBCALL opdis_set_disassembler_options( opdis_t o, const char * options );
 
-// instructions are emitted to display
 /*!
  * \fn opdis_set_display( opdis_t, OPDIS_DISPLAY, void * )
  * \ingroup configuration
+ * \brief Set the callback used to display or store disassembled instructions.
+ * \details This sets the function invoked to manage disassembled instructions.
+ *          The callback is invoked every time an instruction is disassembled.
+ *          It is expected to display or store the instruction, which will be
+ *          discarded when the next address is disassembled. The default
+ *          handler writes to STDOUT.
+ * \param o opdis disassembler to configure.
+ * \param fn The callback function.
+ * \param arg An optional argument to pass to the callback function.
+ * \note If the callback stores the instruction, it must make a copy using
+ *       opdis_insn_dupe().
  */
 void LIBCALL opdis_set_display( opdis_t o, OPDIS_DISPLAY fn, void * arg );
 
-// instructions are emitted to handler; handler makes decision to continue
 /*!
  * \fn opdis_set_handler( opdis_t, OPDIS_HANDLER, void * )
  * \ingroup configuration
+ * \brief Set the callback used to determine whether to continue disassembly.
+ * \details This sets the function invoked to determine whether to continue
+ *          linear or control-flow disassembly after each instruction has
+ *          been disassembled and displayed. The default handler will halt
+ *          disassembly if an invalid instruction has been encountered, or
+ *          (for control-flow) if the end of a control-flow branch (i.e. a
+ *          return statement) has been reached.
+ * \param o opdis disassembler to configure.
+ * \param fn The callback function.
+ * \param arg An optional argument to pass to the callback function.
  */
 void LIBCALL opdis_set_handler( opdis_t o, OPDIS_HANDLER fn, void * arg );
 
-// instructions are submitted to the decoder as an array of strings;
-// the decoder fills the instruction object
 /*!
  * \fn opdis_set_decoder( opdis_t, OPDIS_DECODER )
  * \ingroup configuration
+ * \brief Set the callback used to build an opdis_insn_t from libopcodes data.
+ * \details This sets the function used to fill an opdis_insn_t based on the
+ *          array of strings (stored in an opdis_insn_buf_t) emitted by
+ *          libopcodes. If the default handler does not support the ISA
+ *          being disassembled (i.e. an architecture other than x86 is
+ *          being used), then this must be overridden in order to get
+ *          more than the ASCII version of the instruction in the
+ *          opdis_insn_t.
+ * \param o opdis disassembler to configure.
+ * \param fn The callback function.
  */
 void LIBCALL opdis_set_decoder( opdis_t o, OPDIS_DECODER fn );
 
-// instruction is emitted to resolver; resolver returns jump/call target
 /*!
  * \fn opdis_set_resolver( opdis_t, OPDIS_RESOLVER, void * )
  * \ingroup configuration
+ * \brief Set the callback used to obtain the buffer offset of a branch target.
+ * \details This sets the function used to convert the target of a branch
+ *          to an offset into the buffer. The control flow disassembler will 
+ *          require a resolver in order to disassemble all non-relative branch 
+ *          targets.
+ * \param o opdis disassembler to configure.
+ * \param fn The callback function.
+ * \param arg An optional argument to pass to the callback function.
+ * \note The resolver receives an instruction and is expected to return
+ *       a valid offset or -1 if no address can be determined. The
+ *       resolver can be a subsystem that manages the stack and registers
+ *       in order to provide more complete branch resolution.
  */
 void LIBCALL opdis_set_resolver( opdis_t o, OPDIS_RESOLVER fn, void * arg );
 
 /*!
  * \fn opdis_set_error_reporter( opdis_t, OPDIS_ERROR, void * )
  * \ingroup configuration
+ * \brief Set the callback used to report errors.
+ * \details This sets the function used to report errors encountered during
+ *          disassembly. The default error reporter simply writes to STDERR.
+ * \param o opdis disassembler to configure.
+ * \param fn The callback function.
+ * \param arg An optional argument to pass to the callback function.
  */
 void LIBCALL opdis_set_error_reporter( opdis_t o, OPDIS_ERROR fn, void * arg );
 
-// size of single insn at address
 /*!
  * \fn opdis_disasm_insn_size( opdis_t, opdis_buf_t, opdis_off_t )
  * \ingroup disassembly
+ * \brief Return the size of the instruction at an offset in the buffer.
+ * \param o opdis disassembler
+ * \param buf The buffer to disassemble
+ * \param offset The offset into the buffer to disassemble.
  */
 size_t LIBCALL opdis_disasm_insn_size( opdis_t o, opdis_buf_t buf, 
 				       opdis_off_t offset );
 
-// disasm single insn at address
 /*!
  * \fn opdis_disasm_insn(opdis_t, opdis_buf_t, opdis_off_t, opdis_insn_t * )
  * \ingroup disassembly
+ * \brief Disassemble a single instruction in the buffer
+ * \param o opdis disassembler
+ * \param buf The buffer to disassemble
+ * \param offset The offset into the buffer to disassemble.
+ * \param insn The op_insn_t to fill with the disassembled instruction
  */
-int LIBCALL opdis_disasm_insn( opdis_t o, opdis_buf_t buf, opdis_off_t offset,
-			       opdis_insn_t * insn );
+size_t LIBCALL opdis_disasm_insn( opdis_t o, opdis_buf_t buf, 
+				  opdis_off_t offset, opdis_insn_t * insn );
 
-// disasm serial range of addresses
 /*!
  * \fn opdis_disasm_linear( opdis_t, opdis_buf_t, opdis_off_t, opdis_off_t )
  * \ingroup disassembly
+ * \brief Disassemble a sequence of instructions in order.
+ * \param o opdis disassembler
+ * \param buf The buffer to disassemble
+ * \param offset The offset into the buffer to start disassembly at.
+ * \param length The number of bytes to disassemble.
  */
 int LIBCALL opdis_disasm_linear( opdis_t o, opdis_buf_t buf, opdis_off_t offset,
 				 opdis_off_t length );
@@ -293,12 +367,22 @@ int LIBCALL opdis_disasm_linear( opdis_t o, opdis_buf_t buf, opdis_off_t offset,
 /*!
  * \fn opdis_disasm_cflow( opdis_t, opdis_buf_t, opdis_off_t )
  * \ingroup disassembly
+ * \brief Disassemble a buffer following flow of control.
+ * \param o opdis disassembler
+ * \param buf The buffer to disassemble
+ * \param offset The offset of the entry point in the buffer
  */
 int LIBCALL opdis_disasm_cflow( opdis_t o, opdis_buf_t buf, 
 				opdis_off_t offset );
 
 /*!
  * \fn opdis_error( opdis_t, opdis_error_t, const char * )
+ * \ingroup disassembly
+ * \brief Send an error message to the error reporter/
+ * \details This is used internally and by callbacks to report errors.
+ * \param o opdis disassembler
+ * \param error The error code
+ * \param msg The detailed error message
  */
 void LIBCALL opdis_error( opdis_t o, opdis_error_t error, const char * msg );
